+++ CHANGES.md
@@ -4,8 +4,8 @@ The released versions correspond to PyPi releases.
 ## Version 4.2.0 (as yet unreleased)
 
 #### New Features
-  * add support for the `buffering` parameter in `open` (text line mode not
-    supported) (see [#549](../../issues/549))
+  * add support for the `buffering` parameter in `open` 
+    (see [#549](../../issues/549))
 
 #### Fixes
   * do not truncate file on failed flush